requirements.txt
RPi.GPIO==0.7.1
Pillow==10.1.0
tk==0.1.0
pypylon==3.0.1
numpy==1.26.4
opencv-python==4.9.0.80